Move default bridge test into linux-only file
Becuase I'm about to add tests that use netlink, and the netlink
package breaks compilation under Windows.
